Mythbuntu: Ensure the background image is painted.
Recent changes in master try to minimise the overlap between native Qt
painting (in MythMainWindow) and actual painter rendering (in
MythPainterWindow). The Mythbuntu theme relied on native Qt painting to
render the background image and this is now missing - so ensure it is
painted by adding to the base theme.

There may still be areas (plugins?) where it is not visible.